requirements.txt
pymodbus==2.5.3
pyyaml==6.0
apscheduler==3.9.1
fastapi==0.68.1
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
uvicorn==0.15.0
python-jose==3.3.0
python-multipart==0.0.5
